Add files via upload
## Analysis and Challenges
### Analysis of Outcomes Based on Launch Date
I decided to use a Parent Category filter and set it to Louise’s preference which was theater. I also set the rows to be months to determine the number amount of successful, failed, canceled, and total theater projects. I then used the pivot table data to create the line graph visualization below.
Outcomes_vs_Goals.png

### Analysis of Outcomes Based on Goals
For the analysis of outcomes based on goals, after using the COUNTIFS() formula and transforming projects' success/failure rates into percentages I filtered the data table to not show the “Live” category and there was no existing data of any project being canceled. I then created a line graph visualization that showed the “Outcomes Based on Goals” with the success/failure percentage on the Y-axis and the goal ranges with each row having a $5,000 range. The visualization is shown below: 
Theater_Outcomes_vs_Launch.png

### Challenges and Difficulties Encountered
Some challenges in this project were that there was not a single data entry for canceled projects and that could cause some confusion on whether there was human error involved, especially since there was an entire column with “0” values for it.

## Results
What are two conclusions you can draw about the Outcomes based on the Launch Date?
1. According to the “Theater Outcomes Based on Launch Date” line graph, the best months to launch are May and June to have a higher success rate, while the worst month seems to be December due to being the lowest month of success.
2. There was not any recorded canceled project, at least not yet, there are still 19 live campaigns. 

What can you conclude about the Outcomes based on Goals?
1. According to the “Outcomes Based on Goals” line graph, the best success rate came from goals within the ranges of less than 1$,000 and $35,000 to $44,999. While the worst goal range to have would be between $45,000 to $49,999 due to having the highest failure rate and nearly 0% success rate.

What are some limitations of this dataset?
Some limitations of this dataset are that there weren’t any recorded canceled projects and also we are unable to make conclusions on some of the 19 live theater values, only speculations.

What are some other possible tables and/or graphs that we could create?
Some other tables we could create could be such as the method of campaigns, comparing time ranges in campaigning (successful vs unsuccessful), and a deeper insight on staff pick and spotlight rather than just being a true or false value.